chore: rename netlifyService to netlifyApi
